Tweak income monthly/yearly stuff
I was not actually able to reproduce an error where I could not remove
the extra income sources. I did find other errors with the feature
though, and I have fixed them here.

I removed `monthly_value` and `yearly_value` in favor of just using the
normal `value` field, and I added a `frequency` field.

Do we want to do this approach on the API side @ericmaicon?

This PR is related to #9